Fix: issue 13 -> Added the value of 'noSession' to the "no session available". This was done as the js is set to ignore that value.
Fix: issue 11 -> Change to the showMessageInDialogBox function in
dialogbox.js
Now that function will accept an extra parameter. This parameter will be
a function that is executed after close is pressed.

Change: The $.post of the createSession and
showParticipatingSessionDetails functions in createSession.js and
participatingSessions.js now close loader if an server error happens.