Fix error #41 Changed: the functions in htmlResponseUtil.py now use the utf 8 encoding.
Because this error affected the creation of the svg images and no
solution was found as minidom could not handle the character the check
for valid text in ajaxUpdateGrid and ajaxCreateGrid now checks for the
UnicodeEncodeError error and explicitly informs the user that a
character can not be used correctly.
A secondary error was also found as debug information was being printed.
This happened because there was no check to see if the site was in
debug mode or not.